\name{Rglpk-package}
\alias{Rglpk-package}
\alias{Rglpk}
\docType{package}
\title{R/GNU Linear Programming Kit Interface}
\description{
This package provides a high level access to the GNU Linear Programming
Kit engine.
}
\details{
\tabular{ll}{
Package: \tab Rglpk\cr
Type: \tab Package\cr
Version: \tab 0.1-5\cr
Date: \tab 2007-12-16\cr
License: \tab GPL version 2.1 or newer\cr
}

}
\author{
  Stefan Theussl
  
  Maintainer: Stefan Theussl <stefan.theussl@wu-wien.ac.at>
}
\references{
GNU Linear Programming Kit
}
\keyword{ package }
\seealso{
  \code{\link[glpk:glpk-package]{glpk}}
}
\examples{
}
