\name{Rglpk-package}
\alias{Rglpk-package}
\alias{Rglpk}
\docType{package}
\title{R/GNU Linear Programming Kit Interface}
\description{
This package provides a high level R interface to the GNU Linear Programing Kit
(GLPK - Version 4.24). GLPK is an open source software for solving
large-scale linear programming (LP), mixed integer linear programming
(MILP) and other related problems.
}
\details{
\tabular{ll}{
Package: \tab Rglpk\cr
Type: \tab Package\cr
Version: \tab 0.1-6\cr
Date: \tab 2007-12-18\cr
License: \tab GPL version 2.1 or newer\cr
}

}
\author{
  Stefan Theussl
  
  Maintainer: Stefan Theussl <stefan.theussl@wu-wien.ac.at>
}
\references{
GNU Linear Programming Kit \url{http://www.gnu.org/software/glpk/}
}
\keyword{ package }
\seealso{
  \code{\link[glpk:glpk-package]{glpk}}
}
\examples{
}
